Refactor: split lib.rs into modules
