+++ utils/src/bin/zpaste.rs
@@ -1,5 +1,5 @@
 use std::env;
-use monotextor_utils::ZPaste;
+use monotextor_utils::zpaste::ZPaste;
 
 fn main() {
     let args: Vec<String> = env::args().skip(1).collect();
